# api/openapi.yaml
openapi: 3.1.0
info:
  title: Castflow VOD API
  version: 1.0.0
  description: Self-hosted video-on-demand platform API
servers:
  - url: http://localhost:8080
security:
  - ApiKeyAuth: []
paths:
  /health:
    get:
      summary: Health check
      security: []
      responses:
        "200":
          description: OK
  /api/v1/videos:
    get:
      summary: List videos
      parameters:
        - name: limit
          in: query
          schema: { type: integer, default: 20 }
        - name: offset
          in: query
          schema: { type: integer, default: 0 }
      responses:
        "200":
          description: Paginated video list
    post:
      summary: Upload video (use /videos/upload instead)
  /api/v1/videos/upload:
    post:
      summary: Upload and transcode video
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file: { type: string, format: binary }
                title: { type: string }
                description: { type: string }
      responses:
        "202":
          description: Upload accepted
  /api/v1/videos/{id}:
    get:
      summary: Get video metadata
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Video details
    delete:
      summary: Delete video
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "204":
          description: Deleted
  /api/v1/videos/{id}/status:
    get:
      summary: Processing status
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Status
  /api/v1/videos/{id}/links:
    get:
      summary: All playback URLs (HLS, DASH, player, iframe, ...)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Playback links
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key